refactor and add simple position editor
